sustindex-/requirements.txt
Django==5.0.6
gunicorn==21.2.0
psycopg[binary]==3.2.3
dj-database-url==2.1.0
whitenoise==6.6.0
django-ckeditor==6.7.0
django-import-export==3.3.9
django-simple-history==3.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==11.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
